Expose `lib/` and remove CommonJS stuff
CommonJS stuff isn't needed for our ingestion of the package, so
removing it lessens the number of files and build steps a little bit.

Although it's sort of bad practice, I've exposed (un-gitignored) the
`lib/` directory to make it easier to use this package with `github:
repo` syntax. The pre-existing logic assumed the package would only be
used from the `pnpm publish` bundle, and by sharing `lib/` we're doing
something along those lines.