Un-gitignore lib/ and export these files explicitly (#5)
Although it's sort of bad practice, I've exposed (un-gitignored) the
`lib/` directory to make it easier to use this package with the `github:
repo` syntax. The previous logic assumed the package would only be used
from the `pnpm publish` bundle (which is created with contents of
`lib/`), so by exposing `lib/` we're effectively making the repo appear
as a published bundle.